Renamed AppointmentActionVaos to AppointmentAction and deleted old AppointmentAction.
